Fix for Issue #1
Get Time in microseconds from timems function. Will ensure better accuracy of rngSeed.

The / 1000 was sometimes negating the added benefit of end.tv_usec. On linux system this was sometimes causing issues when the end.tv_usec value was in the 1e3 or 1e4 range.

You would have to catch it at the precise time for you to reproduce the issue. Multiple starts are required usually.